requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
httpx>=0.27
pypdf>=4.0
python-multipart>=0.0.9
python-dotenv>=1.0
